Remove old compaction job creation lambda
